# server/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Distributed arithmetic expressions calculation system",
        "title": "Web Calculator API",
        "contact": {
            "name": "API Support",
            "url": "http://localhost:8080",
            "email": "support@calculator.ru"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/calculate": {
            "post": {
                "description": "Add new arithmetic expression for calculation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "expressions"
                ],
                "summary": "Create new expression",
                "parameters": [
                    {
                        "description": "Expression data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CalculateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.CalculateResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/models.ErrResponse"
                        }
                    }
                }
            }
        },
        "/expressions": {
            "get": {
                "description": "Get list of all expressions with their statuses",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "expressions"
                ],
                "summary": "Get all expressions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ExpressionsResponse"
                        }
                    }
                }
            }
        },
        "/expressions/{id}": {
            "get": {
                "description": "Get expression details by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "expressions"
                ],
                "summary": "Get expression by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Expression ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Expression"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.CalculateRequest": {
            "type": "object",
            "properties": {
                "expression": {
                    "type": "string",
                    "example": "2+2*2"
                }
            }
        },
        "models.CalculateResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "models.ErrResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "models.Expression": {
            "type": "object",
            "properties": {
                "expression": {
                    "type": "string",
                    "example": "2+2*2"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "result": {
                    "type": "string",
                    "example": "6"
                },
                "status": {
                    "type": "string",
                    "example": "completed"
                }
            }
        },
        "models.ExpressionsResponse": {
            "type": "object",
            "properties": {
                "expressions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Expression"
                    }
                }
            }
        }
    }
}